% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_testplan.R
\name{build_testplan}
\alias{build_testplan}
\title{build_testplan}
\usage{
build_testplan(pkgdir, deps = FALSE)
}
\arguments{
\item{pkgdir}{PARAM_DESCRIPTION}

\item{deps}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
#EXAMPLE1
 }
}
\author{
Lorenzo Busetto, phD (2017) <lbusett@gmail.com>
}
